requirements.txt
Django>=3.0
psycopg2>=2.8,<3.0
django-fa >= 1.0
django-lint >= 2.0
autopep8 >= 1.5
pylint >= 2.4
pylint-django >= 2.0
Pillow >= 7.0
django-autocomplete-light >= 3.5
djangorestframework >= 3.11